+++ chat-components/src/components/loadingpane/common/defaultProps/defaultStyles/defaultLoadingPaneIconImageProps.ts
@@ -4,7 +4,7 @@ import { ModernChatIconBase64 } from "../../../../../assets/Icons";
 
 export const defaultLoadingPaneIconImageProps: IImageProps = {
     src: ModernChatIconBase64,
-    imageFit: ImageFit.center,
+    imageFit: ImageFit.centerContain,
     width: "86px",
     height: "86px",
     shouldFadeIn: false,